removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5198 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIDetectorWrapper.cpp
@@ -53,7 +53,7 @@ GUIDetectorWrapper::GUIDetectorWrapper(GUIGlObjectStorage &idStorage,
 
 GUIDetectorWrapper::GUIDetectorWrapper(GUIGlObjectStorage &idStorage,
                                        std::string id,
-                                       size_t glID) throw()
+                                       GLuint glID) throw()
         : GUIGlObject_AbstractAdd(idStorage, id, glID, GLO_DETECTOR)
 {}
